appease the sarah linter
