+++ cmd/doc.go
@@ -1,0 +1,2 @@
+// Package cmd is our cobra/viper cli implementation
+package cmd
