+++ config/.env-dev.example
@@ -42,3 +42,8 @@ DATUM_SESSIONS_ENCRYPTION_KEY=IF841pLk48ibOVo6w9UIkr9OBCkjXG2q
 DATUM_EMAIL_SEND_GRID_API_KEY=SG.FakeAPIKey
 DATUM_EMAIL_TESTING=true
 DATUM_EMAIL_URL_BASE=http://localhost:17608/
+
+# Sentry Settings
+DATUM_SENTRY_ENABLED=
+DATUM_SENTRY_DSN=
+DATUM_SENTRY_ENVIRONMENT=
