+++ config/.env.example
@@ -106,3 +106,7 @@ DATUM_TRACING_OTLP_HEADERS=
 DATUM_TRACING_OTLP_COMPRESSION=
 DATUM_TRACING_OTLP_TIMEOUT=
 
+# Sentry Settings
+DATUM_SENTRY_ENABLED=
+DATUM_SENTRY_DSN=
+DATUM_SENTRY_ENVIRONMENT=
