+++ internal/utils/sentry/doc.go
@@ -1,0 +1,2 @@
+// Package sentry is a helper package for sentry integration
+package sentry
